Refactoring Mocks
